fix(core): preserve traversal compatibility
